+++ gulp-tasks/watch.js
@@ -33,17 +33,8 @@ module.exports = {
 			`${ pkg.square1.paths.core_theme_components }**/*.pcss`,
 			`${ pkg.square1.paths.core_theme_routes }**/*.pcss`,
 			`${ pkg.square1.paths.core_theme_integrations }**/*.pcss`,
-			`!${ pkg.square1.paths.core_theme_pcss }legacy.pcss`,
-			`!${ pkg.square1.paths.core_theme_pcss }content/page/_legacy.pcss`,
 			`!${ pkg.square1.paths.core_admin_pcss }**/*.pcss`,
 		], gulp.parallel( 'postcss:theme' ) );
-
-		// watch the legacy postcss
-
-		gulp.watch( [
-			`${ pkg.square1.paths.core_theme_pcss }legacy.pcss`,
-			`${ pkg.square1.paths.core_theme_routes }unsupported_browser/css/unsupported-browser.pcss`,
-		], gulp.parallel( 'postcss:themeLegacy' ) );
 
 		// watch the login postcss
 
@@ -59,8 +50,6 @@ module.exports = {
 			`${ pkg.square1.paths.core_theme_routes }**/*.pcss`,
 			//`${ pkg.square1.paths.core_theme_integrations }**/*.pcss`,
 			`${ pkg.square1.paths.core_admin_pcss }mce-editor.pcss`,
-			`!${ pkg.square1.paths.core_theme_pcss }legacy.pcss`,
-			`!${ pkg.square1.paths.core_theme_pcss }content/page/_legacy.pcss`,
 		], gulp.parallel( 'postcss:adminMCEEditor' ) );
 
 		// watch the block editor styles postcss
@@ -72,8 +61,6 @@ module.exports = {
 			//`${ pkg.square1.paths.core_theme_integrations }**/*.pcss`,
 			`${ pkg.square1.paths.core_admin_pcss }block-editor.pcss`,
 			`${ pkg.square1.paths.core_admin_pcss }block-editor/**/*.pcss`,
-			`!${ pkg.square1.paths.core_theme_pcss }legacy.pcss`,
-			`!${ pkg.square1.paths.core_theme_pcss }content/page/_legacy.pcss`,
 		], gulp.parallel( 'postcss:adminBlockEditor' ) );
 
 		// watch the admin styles postcss